Extact polling logic to custom hook
